% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\alias{fetch_dependencies}
\alias{check_dependencies}
\title{ariseng dependencies}
\usage{
fetch_dependencies()

check_dependencies(silent = FALSE)
}
\arguments{
\item{silent}{logical; whether or not to provide informative messages.
Default is \code{FALSE}.}
}
\value{
\code{TRUE}/\code{FALSE} (invisibly).
}
\description{
Download and install packages required to use this project.

Check whether dependencies are available in R.
}
\details{
The function carries out a check as to whether the packages already
exist for the current version of R. The packages are:
\itemize{
  \item \code{gWidgets2}
  \item \code{gWidgets2RGtk2}
  \item \code{RGtk2}
  \item \code{DataEntry}
}

The packages 'gWidgets2', 'gWidgets2RGtk2', 'RGtk2', 'DataEntry'
are needed for this package to work. In carrying out the check, informative
message can optionally be provided by setting \code{silent = TRUE}.
}
