% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fall_back.R
\name{fall_back}
\alias{fall_back}
\title{Fall-back function}
\usage{
fall_back()
}
\description{
Use spreadsheet format for data entry instead of GUI.
}
\details{
A fall-back function to use for data entry in the event of failure to
to initialize the GUI
}
