% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{fetch_settings}
\alias{fetch_settings}
\title{Fetch the project template/data}
\usage{
fetch_settings(path = ".")
}
\arguments{
\item{path}{The path to a directory where the settings are to be
saved.}
}
\description{
Enables the user to use the predefined variables for this project
}
