% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_project.R
\name{reset_project}
\alias{reset_project}
\title{Reset the Project}
\usage{
reset_project(proj)
}
\arguments{
\item{proj}{The path to the project file (with extension \emph{.dte}).}
}
\description{
Starting the entire project from scratch
}
