% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_data}
\alias{save_data}
\alias{save_data_sql}
\title{Save Data from Project Files}
\usage{
save_data_sql(dte.file, dest = getwd())
}
\arguments{
\item{dte.file}{Character vector of length 1 - path to the project file}

\item{dest}{The destination folder. Defaults to working directory.}
}
\description{
Save Data from Project Files
}
\details{
\emph{dte.file} is the \code{.dte} project file specific to the
project.

\emph{dest} is the directory where the database is to be saved.
}
